Generate payment_id instead of using identity
This now matches what the DCT was asking for, in the first place.
This is another case where I had been using an int, instead of a varchar.  I had only been looking at rental so far.  Because citation and other systems handle payments, I had to change what I was using as the payment_id.