Create a bond permit for each contact from ROW
This creates a bond permit for all the unique principal contacts from ROW.  If
there is no company_id, then the contact_id is used as the principal.

This permit will be associated with the principal contact and will have the
bonds associated with it.  Then, as the principal applies for various permits,
we will associate their ROW permits with this "bond permit".

Updates #42