Create status strings that reflect status+active
Most of the legacy applications use statuses that do not reflect whether the
record is still active or not.  Rentpro and Citiation have a seperate
active/inactive flag.

In the mapping spreadsheet, we need to mark whether the status is a active or
not.  Only the records with an active type of status will have their workflows
generated.

Updates #54