Rental: link inspection cases to permits
I no longer create inspection cases for inactive rental properties.  This is
because there is no way to know what end date to use for the inspection.
Without an end date, I'm not certain that the inspection_case won't be
scheduling new inspections.

To avoid this, I just don't create inspection_cases for inactive properties.

Updates #57